Testing/requirements.txt
google-generativeai
PyMuPDF
pdf2image
Pillow
json5
